Move feature_converters and evaluator into seqio.
PiperOrigin-RevId: 349458399
